wallet: broken
The test fails because the address of the two balances are different.
When we change the value of the balance in code, we are working with a
copy of the balance from test. In Go, **when you call a function or a
method the arguments are copied**.

```
➜ gopher_zone/06_pointers_and_errors main ✗ (?) go test -cover -v
=== RUN   TestWallet
address of balance in deposit is 0x140000a00b0
address of balance in test is 0x140000a00a8
    wallet_test.go:18: got 0.00 want 10.00
--- FAIL: TestWallet (0.00s)
FAIL
coverage: 100.0% of statements
exit status 1
FAIL	gopher_zone/06_pointers_and_errors	1.042s
```

This can be fixed with _pointers_.