store: update happy path to ensure it is not cancelled
`context` has a method `Done()` which returns a channel which gets sent
a signal when the context is done or cancelled. we listen on that signal
and call `store.Cancel` if we get it but ignore it otherwise if our
store manages to `Fetch` before it.

this is possible if we run `Fetch` in a goroutine and it will write the
result into a new channel `data`. we then use `select` to effectively
race the two async processes and then we either write a `response` or
cancel.